Make NaN_val const, declare the saved state of the FPU only for the mathematical library in use, and make three conversions explicit, which GCC and Clang warned about in the autotools and meson builds
NaN_val, defined in gaol_port.h, was an unused variable in most of the files
including it (-Wunused-variable), and save_fpu_state_crlibm in gaol_common.cpp
with mathlib. The conversions are the double midpoint of an interval of floats
(gaol_intervalf.cpp), the table of MSB_position() (gaol_interval_sse.cpp) and
the seed of srand() (check/performances.cpp).

+++ check/performances.cpp
@@ -74,7 +74,7 @@ int main(int argc, char *argv[])
 
 	string fun(argv[1]);
 
-	srand(seed);
+	srand(static_cast<unsigned int>(seed));
 	srand48(seed);
 
 	double prob[4] = { pdenorm, // denormal
