+++ gaol/gaol_intervalf.cpp
@@ -443,7 +443,7 @@ namespace gaol {
     }
 
     GAOL_RND_ENTER();
-    float middle = left()+(.5*right() - .5*left()); 
+    float middle = static_cast<float>(left()+(.5*right() - .5*left()));
     GAOL_RND_LEAVE();
     return middle;
   }
